Move createAccount helper to persistence and use getHermezAccounts abstraction everywhere
